+++ test/models/sheet_variable_test.rb
@@ -73,9 +73,9 @@ class SheetVariableTest < ActiveSupport::TestCase
   end
 
   test 'get time of day' do
-    assert_equal Formatters::TimeFormatter, Formatters.for(sheet_variables(:time).variable).class
-    assert_equal '22:30:00', sheet_variables(:time).get_response(:name)
-    assert_equal '81000', sheet_variables(:time).get_response(:raw)
+    assert_equal Formatters::TimeOfDayFormatter, Formatters.for(sheet_variables(:time_of_day).variable).class
+    assert_equal '22:30:00', sheet_variables(:time_of_day).get_response(:name)
+    assert_equal '81000', sheet_variables(:time_of_day).get_response(:raw)
   end
 
   test 'get time duration' do